releasetools: Allow for other pre-built images to be added to the build
Allow for other pre-built images to be added to the build system.

The current build system allows for a pre-built vendor.img to be included as
part of the breakfast. However while other partitions, like firmware or others,
can be included in the AB_OTA_PARTITIONS list, they will not be included in the
OTA file as they do get copied to the right output directory during the build.

This patch follows the same process for copying in vendor.img to the output
directory, but allows you to define more than one via the following settings:

    BOARD_PREBUILT_IMAGES
    BOARD_PREBUILT_IMAGEPATH

BOARD_PREBUILT_IMAGES is a list of partition names to copy over, for example:

    BOARD_PREBUILT_IMAGES := bluetooth cmnlib modem

Where as BOARD_PREBUILT_IMAGEPATH is the absolute path location where the
image files can be found.  For example:

    BOARD_PREBUILT_IMAGEPATH := /home/user/vendor/images

Used in conjunction with change 277134, this allows a full OTA file to be built
with additional pre-built images.

OG-Change-Id: I411366ac9c0d9ff7b155bfb670b8be4d6649f712
Change-Id: I512212cfd20f97e9bea7d406a993d1b97ccbf28a